auth-service/src/main/java/com/money/management/auth/security/impl/AuthenticationManagerServiceImpl.java
package com.money.management.auth.security.impl;

import com.money.management.auth.payload.LoginRequest;
import com.money.management.auth.security.AuthenticationManagerService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Service;

@Service
public class AuthenticationManagerServiceImpl implements AuthenticationManagerService {

    private AuthenticationManager authenticationManager;

    @Autowired
    public AuthenticationManagerServiceImpl(AuthenticationManager authenticationManager) {
        this.authenticationManager = authenticationManager;
    }

    @Override
    public Authentication authenticate(LoginRequest loginRequest) {
        UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
                loginRequest.getEmail(),
                loginRequest.getPassword()
        );

        Authentication authentication = authenticationManager.authenticate(authenticationToken);

        SecurityContextHolder.getContext().setAuthentication(authentication);

        return authentication;
    }
}
